requirements.txt
alembic==1.7.1
asgiref==3.4.1
click==8.0.1
fastapi==0.68.1
greenlet==1.1.1
h11==0.12.0
httptools==0.2.0
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2==2.9.1
pydantic==1.8.2
python-dotenv==0.19.0
python-slugify==5.0.2
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
text-unidecode==1.3
typing-extensions==3.10.0.2
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==9.1
